% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{applyFunctionToResultsDirectories}
\alias{applyFunctionToResultsDirectories}
\title{Loop a function over a set of results directories}
\usage{
applyFunctionToResultsDirectories(fun, maindir = "./", pattern = "", ...)
}
\arguments{
\item{fun}{the function to apply}

\item{maindir}{the parent directory}
}
\description{
Loop a function over a set of results directories
}
