% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{calcHybridForce}
\alias{calcHybridForce}
\title{Generate the hybrid force functions given peak, location of peak, and cutoff}
\usage{
calcHybridForce(x = seq(-0.3, 2.5, 0.01), max_F = 0.2, peak_location = 1.1)
}
\arguments{
\item{x}{the x locations}

\item{max_F}{the peak force}

\item{peak_location}{the peak location}
}
\description{
Generate the hybrid force functions given peak, location of peak, and cutoff
}
