% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{convertNamedTimeStepVector}
\alias{convertNamedTimeStepVector}
\title{Convert a named vector from time steps to data frame with days}
\usage{
convertNamedTimeStepVector(v, value_name = "v")
}
\arguments{
\item{v}{the named vector}

\item{value_name}{the string part of the names to gsub out}
}
\description{
Convert a named vector from time steps to data frame with days
}
