% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{getCellAncestory}
\alias{getCellAncestory}
\title{Load in a results.vizancestors file
Reading in of the cell ancestory output}
\usage{
getCellAncestory(filepath)
}
\arguments{
\item{filepath}{the path to the containing results folder}
}
\description{
Load in a results.vizancestors file
Reading in of the cell ancestory output
}
