% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{getCellCounts}
\alias{getCellCounts}
\title{Calculate the counts of a list of time step data frames}
\usage{
getCellCounts(data, column)
}
\arguments{
\item{data}{a list of the chaste results data at each time step}

\item{column}{the column of the data frame to count over}
}
\value{
a data frame containing the counts
export
}
\description{
Calculate the counts of a list of time step data frames
}
