% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{getVoronoiConnectivity2d}
\alias{getVoronoiConnectivity2d}
\title{Approximate connectivity by voronoi tesselation in 2d}
\usage{
getVoronoiConnectivity2d(x, y, width, b_dist = 2)
}
\arguments{
\item{x}{the x locations}

\item{y}{the y locations}

\item{width}{the periodic width}

\item{b_dist}{the boundary distance}
}
\description{
Approximate connectivity by voronoi tesselation in 2d
}
