% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{getVoronoiConnectivity3d}
\alias{getVoronoiConnectivity3d}
\title{Approximate connectivity by voronoi tesselation}
\usage{
getVoronoiConnectivity3d(x, y, z, width, b_dist = 2)
}
\arguments{
\item{x}{the x locations}

\item{y}{the y locations}

\item{z}{the z locations}

\item{width}{the periodic width}

\item{b_dist}{the boundary distance}
}
\description{
Approximate connectivity by voronoi tesselation
}
