% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readCellData}
\alias{readCellData}
\title{Read a cell data file}
\usage{
readCellData(filepath, var_name, dimensions = 2)
}
\arguments{
\item{var_name}{the variable name of the cell data from chaste}
}
\value{
A data frame containing the cell data output
}
\description{
Loads in data from a cell data output
}
