% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readCellVelocities}
\alias{readCellVelocities}
\title{Read a cell velocities file}
\usage{
readCellVelocities(filepath, dimensions = 3)
}
\arguments{
\item{filepath}{the folder path to where the simulation output is located}

\item{dimensions}{the number of dimensions of the simulation}
}
\description{
Loads in data from NodeVelocityWriter output
Note, has got my modification where the cell id is not repeated
}
