% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readChasteResultsFile}
\alias{readChasteResultsFile}
\title{Efficient loading of a Chaste results file}
\usage{
readChasteResultsFile(filename, filepath = ".", columns = "v")
}
\arguments{
\item{filename}{the name or filepath of the results file}

\item{filepath}{the path to the file}

\item{columns}{(optional) a vector of names for the columns of the outputs for each time step}
}
\value{
A list of data frames, one for each time step. If columns are specified data is split into these columns.
}
\description{
Reading in of the generic results.viz[data] file with timesteps
}
